refactoring and removing legacy code
