+++ src/f3dasm/_src/design/domain.py
@@ -11,16 +11,10 @@ from __future__ import annotations
 # Standard
 import math
 import pickle
-import sys
 from dataclasses import dataclass, field
 from pathlib import Path
-from typing import (Any, Dict, Iterable, Iterator, List, Optional, Sequence,
-                    Type)
-
-if sys.version_info < (3, 8):  # NOQA
-    from typing_extensions import Literal  # NOQA
-else:
-    from typing import Literal
+from typing import (Any, Dict, Iterable, Iterator, List, Literal, Optional,
+                    Sequence, Type)
 
 # Third-party core
 import numpy as np
