+++ src/f3dasm/_src/optimization/numpy_implementations.py
@@ -1,5 +1,5 @@
 """
-Random Search optimizer
+Optimizers based from the numpy library
 """
 
 #                                                                       Modules
