+++ src/f3dasm/_src/optimization/scipy_implementations.py
@@ -1,5 +1,5 @@
 """
-L-BFGS-B optimizer
+Optimizers based from the scipy.optimize library
 """
 
 #                                                                       Modules
@@ -9,8 +9,8 @@ L-BFGS-B optimizer
 from dataclasses import dataclass
 from typing import List
 
-from .adapters.scipy_implementations import _SciPyOptimizer
 # Locals
+from .adapters.scipy_implementations import _SciPyOptimizer
 from .optimizer import OptimizerParameters
 
 #                                                          Authorship & Credits
@@ -20,6 +20,25 @@ __credits__ = ['Martin van der Schelling']
 __status__ = 'Stable'
 # =============================================================================
 #
+# =============================================================================
+
+
+@dataclass
+class CG_Parameters(OptimizerParameters):
+    """CG Parameters"""
+
+    gtol: float = 0.0
+
+
+class CG(_SciPyOptimizer):
+    """CG"""
+    require_gradients: bool = True
+    method: str = "CG"
+    hyperparameters: CG_Parameters = CG_Parameters()
+
+    def get_info(self) -> List[str]:
+        return ['Stable', 'First-Order', 'Single-Solution']
+
 # =============================================================================
 
 
@@ -39,3 +58,24 @@ class LBFGSB(_SciPyOptimizer):
 
     def get_info(self) -> List[str]:
         return ['Stable', 'First-Order', 'Single-Solution']
+
+# =============================================================================
+
+
+@dataclass
+class NelderMead_Parameters(OptimizerParameters):
+    """Hyperparameters for NelderMead optimizer"""
+
+    xatol: float = 0.0
+    fatol: float = 0.0
+    adaptive: bool = False
+
+
+class NelderMead(_SciPyOptimizer):
+    """Nelder-Mead"""
+    require_gradients: bool = False
+    method: str = "Nelder-Mead"
+    hyperparameters: NelderMead_Parameters = NelderMead_Parameters()
+
+    def get_info(self) -> List[str]:
+        return ['Fast', 'Global', 'First-Order', 'Single-Solution']
