+++ src/f3dasm/design.py
@@ -6,11 +6,7 @@ Module for design-of-experiments
 
 # Local
 from ._src.design.domain import Domain, make_nd_continuous_domain
-from ._src.design.parameter import (PARAMETERS, _CategoricalParameter,
-                                    _ConstantParameter, _ContinuousParameter,
-                                    _DiscreteParameter, _Parameter)
-from ._src.experimentdata._data import _Data
-from ._src.experimentdata._jobqueue import NoOpenJobsError, Status, _JobQueue
+from ._src.experimentdata._jobqueue import NoOpenJobsError, Status
 
 #                                                          Authorship & Credits
 # =============================================================================
@@ -22,16 +18,8 @@ __status__ = 'Stable'
 # =============================================================================
 
 __all__ = [
-    '_CategoricalParameter',
-    '_ConstantParameter',
-    '_ContinuousParameter',
-    '_DiscreteParameter',
     'Domain',
     'make_nd_continuous_domain',
     'NoOpenJobsError',
-    'PARAMETERS',
-    '_Parameter',
     'Status',
-    '_Data',
-    '_JobQueue',
 ]
