+++ tests/design/test_designofexperiments.py
@@ -2,8 +2,10 @@ import numpy as np
 import pandas as pd
 import pytest
 
-from f3dasm.design import (Domain, _CategoricalParameter, _ContinuousParameter,
-                           _DiscreteParameter, make_nd_continuous_domain)
+from f3dasm._src.design.parameter import (_CategoricalParameter,
+                                          _ContinuousParameter,
+                                          _DiscreteParameter)
+from f3dasm.design import Domain, make_nd_continuous_domain
 
 pytestmark = pytest.mark.smoke
